leadscore-api-py/requirements.txt
fastapi
uvicorn[standard]
google-genai
cloud-sql-python-connector
pg8000
sqlalchemy
pydantic
python-json-logger
google-cloud-logging
google-cloud-error-reporting
